Updates tests
